--- src/components/file-browser/file-browser-types.ts
@@ -1,11 +1,0 @@
-export interface FileBrowserFile {
-  path: string
-  url: string
-  size: number
-  createdAt: number
-  updatedAt: number
-}
-
-export interface FileBrowserDirectory {
-  path: string
-}
