+++ src/libs/ajax/IncrementalResponse.ts
@@ -1,0 +1,5 @@
+export default interface IncrementalResponse<T> {
+  results: T[]
+  getNextPage(options?: { signal?: AbortSignal }): Promise<IncrementalResponse<T>>
+  hasNextPage: boolean
+}
